requirements.txt
asgiref==3.11.0
colorama==0.4.6
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
